+++ src/index.html
@@ -17,6 +17,6 @@
   
 </body>
 <script type="module" src="main.js"></script>
-<script type="module" src="database.js"></script>
+<script type="module" src="firebase.js"></script>
   
 </html>
